+++ src/promptflow/promptflow/_core/tool_meta_generator.py
@@ -114,11 +114,11 @@ def collect_tool_functions_in_module(m):
 
 
 def collect_flow_entry_in_module(m, entry):
-    entry = entry.split(":")[-1]
-    func = getattr(m, entry, None)
+    func_name = entry.split(":")[-1]
+    func = getattr(m, func_name, None)
     if isinstance(func, types.FunctionType):
         return func
-    raise PythonParsingError(
+    raise PythonLoadError(
         message_format="Failed to collect flow entry '{entry}' in module '{module}'.",
         entry=entry,
         module=m.__name__,
@@ -253,6 +253,21 @@ def load_python_module_from_file(src_file: Path):
             error_type_and_message=error_type_and_message,
         ) from e
     return m
+
+
+def load_python_module_from_entry(entry: str):
+    try:
+        module_name = entry.split(":")[0]
+        return importlib.import_module(module_name)
+    except Exception as e:
+        error_type_and_message = f"({e.__class__.__name__}) {e}"
+        raise PythonLoadError(
+            message_format="Failed to load python module '{module_name}' from entry '{entry}': "
+            "{error_type_and_message}",
+            module_name=module_name,
+            entry=entry,
+            error_type_and_message=error_type_and_message,
+        ) from e
 
 
 def load_python_module(content, source=None):
@@ -477,8 +492,12 @@ def generate_tool_meta_in_subprocess(
     return tool_dict, exception_dict
 
 
-def generate_flow_meta_dict_by_file(path: str, entry: str, source: str = None):
-    m = load_python_module_from_file(Path(path))
+def generate_flow_meta_dict_by_file(entry: str, source: str = None, path: str = None):
+    if path:
+        m = load_python_module_from_file(Path(path))
+    else:
+        m = load_python_module_from_entry(entry)
+
     f = collect_flow_entry_in_module(m, entry)
     # Since the flow meta is generated from the entry function, we leverage the function
     # _parse_tool_from_function to parse the interface of the entry function to get the inputs and outputs.
@@ -492,6 +511,8 @@ def generate_flow_meta_dict_by_file(path: str, entry: str, source: str = None):
         for k, v in tool.inputs.items():
             # We didn't support specifying multiple types for inputs, so we only take the first one.
             flow_meta["inputs"][k] = {"type": v.type[0].value}
+            if v.default is not None:
+                flow_meta["inputs"][k]["default"] = v.default
     if tool.outputs:
         flow_meta["outputs"] = {}
         for k, v in tool.outputs.items():