+++ src/promptflow/promptflow/_sdk/_service/apis/collector.py
@@ -8,10 +8,12 @@
 # to provide OTLP/HTTP endpoint as OTEL collector
 
 import json
+import logging
 import traceback
 from datetime import datetime
+from typing import Callable, Optional
 
-from flask import current_app, request
+from flask import request
 from google.protobuf.json_format import MessageToJson
 from opentelemetry.proto.collector.trace.v1.trace_service_pb2 import ExportTraceServiceRequest
 
@@ -21,12 +23,32 @@ from promptflow._constants import (
     SpanResourceAttributesFieldName,
     SpanResourceFieldName,
 )
+from promptflow._sdk._constants import TRACE_DEFAULT_SESSION_ID
 from promptflow._sdk._utils import parse_kv_from_pb_attribute
 from promptflow._sdk.entities._trace import Span
 from promptflow._utils.thread_utils import ThreadWithContextVars
 
 
-def trace_collector():
+def trace_collector(
+    get_created_by_info_with_cache: Callable,
+    logger: logging.Logger,
+    cloud_trace_only: bool = False,
+    credential: Optional[object] = None,
+):
+    """Collect traces from OTLP/HTTP endpoint and write to local/remote storage.
+
+    This function is target to be reused in other places, so pass in get_created_by_info_with_cache and logger to avoid
+    app related dependencies.
+
+    :param get_created_by_info_with_cache: A function that retrieves information about the creator of the trace.
+    :type get_created_by_info_with_cache: Callable
+    :param logger: The logger object used for logging.
+    :type logger: logging.Logger
+    :param cloud_trace_only: If True, only write trace to cosmosdb and skip local trace. Default is False.
+    :type cloud_trace_only: bool
+    :param credential: The credential object used to authenticate with cosmosdb. Default is None.
+    :type credential: Optional[object]
+    """
     content_type = request.headers.get("Content-Type")
     # binary protobuf encoding
     if "application/x-protobuf" in content_type:
@@ -39,6 +61,8 @@ def trace_collector():
                 attribute_dict = json.loads(MessageToJson(attribute))
                 attr_key, attr_value = parse_kv_from_pb_attribute(attribute_dict)
                 resource_attributes[attr_key] = attr_value
+            if SpanResourceAttributesFieldName.SESSION_ID not in resource_attributes:
+                resource_attributes[SpanResourceAttributesFieldName.SESSION_ID] = TRACE_DEFAULT_SESSION_ID
             resource = {
                 SpanResourceFieldName.ATTRIBUTES: resource_attributes,
                 SpanResourceFieldName.SCHEMA_URL: resource_span.schema_url,
@@ -47,11 +71,19 @@ def trace_collector():
                 for span in scope_span.spans:
                     # TODO: persist with batch
                     span = Span._from_protobuf_object(span, resource=resource)
-                    span._persist()
+                    if not cloud_trace_only:
+                        span._persist()
                     all_spans.append(span)
 
-        # Create a new thread to write trace to cosmosdb to avoid blocking the main thread
-        ThreadWithContextVars(target=_try_write_trace_to_cosmosdb, args=(all_spans,)).start()
+        if cloud_trace_only:
+            # If we only trace to cloud, we should make sure the data writing is success before return.
+            _try_write_trace_to_cosmosdb(all_spans, get_created_by_info_with_cache, logger, credential)
+        else:
+            # Create a new thread to write trace to cosmosdb to avoid blocking the main thread
+            ThreadWithContextVars(
+                target=_try_write_trace_to_cosmosdb,
+                args=(all_spans, get_created_by_info_with_cache, logger, credential),
+            ).start()
         return "Traces received", 200
 
     # JSON protobuf encoding
@@ -59,7 +91,9 @@ def trace_collector():
         raise NotImplementedError
 
 
-def _try_write_trace_to_cosmosdb(all_spans):
+def _try_write_trace_to_cosmosdb(
+    all_spans, get_created_by_info_with_cache: Callable, logger: logging.Logger, credential: Optional[object] = None
+):
     if not all_spans:
         return
     try:
@@ -69,37 +103,50 @@ def _try_write_trace_to_cosmosdb(all_spans):
         resource_group_name = resource_attributes.get(SpanResourceAttributesFieldName.RESOURCE_GROUP_NAME, None)
         workspace_name = resource_attributes.get(SpanResourceAttributesFieldName.WORKSPACE_NAME, None)
         if subscription_id is None or resource_group_name is None or workspace_name is None:
-            current_app.logger.debug("Cannot find workspace info in span resource, skip writing trace to cosmosdb.")
+            logger.debug("Cannot find workspace info in span resource, skip writing trace to cosmosdb.")
             return
 
-        current_app.logger.info(f"Start writing trace to cosmosdb, total spans count: {len(all_spans)}.")
+        logger.info(f"Start writing trace to cosmosdb, total spans count: {len(all_spans)}.")
         start_time = datetime.now()
-        from promptflow._sdk._service.app import CREATED_BY_FOR_LOCAL_TO_CLOUD_TRACE
         from promptflow.azure._storage.cosmosdb.client import get_client
         from promptflow.azure._storage.cosmosdb.span import Span as SpanCosmosDB
         from promptflow.azure._storage.cosmosdb.summary import Summary
 
         # Load span and summary clients first time may slow.
         # So, we load 2 client in parallel for warm up.
-        span_thread = ThreadWithContextVars(
-            target=get_client, args=(CosmosDBContainerName.SPAN, subscription_id, resource_group_name, workspace_name)
+        span_client_thread = ThreadWithContextVars(
+            target=get_client,
+            args=(CosmosDBContainerName.SPAN, subscription_id, resource_group_name, workspace_name, credential),
         )
-        span_thread.start()
+        span_client_thread.start()
 
-        get_client(CosmosDBContainerName.LINE_SUMMARY, subscription_id, resource_group_name, workspace_name)
+        # Load created_by info first time may slow. So, we load it in parallel for warm up.
+        created_by_thread = ThreadWithContextVars(target=get_created_by_info_with_cache)
+        created_by_thread.start()
 
-        span_thread.join()
+        get_client(CosmosDBContainerName.LINE_SUMMARY, subscription_id, resource_group_name, workspace_name, credential)
+
+        span_client_thread.join()
+        created_by_thread.join()
+
+        created_by = get_created_by_info_with_cache()
 
         for span in all_spans:
-            span_client = get_client(CosmosDBContainerName.SPAN, subscription_id, resource_group_name, workspace_name)
-            result = SpanCosmosDB(span, CREATED_BY_FOR_LOCAL_TO_CLOUD_TRACE).persist(span_client)
+            span_client = get_client(
+                CosmosDBContainerName.SPAN, subscription_id, resource_group_name, workspace_name, credential
+            )
+            result = SpanCosmosDB(span, created_by).persist(span_client)
             # None means the span already exists, then we don't need to persist the summary also.
             if result is not None:
                 line_summary_client = get_client(
-                    CosmosDBContainerName.LINE_SUMMARY, subscription_id, resource_group_name, workspace_name
+                    CosmosDBContainerName.LINE_SUMMARY,
+                    subscription_id,
+                    resource_group_name,
+                    workspace_name,
+                    credential,
                 )
-                Summary(span, CREATED_BY_FOR_LOCAL_TO_CLOUD_TRACE).persist(line_summary_client)
-        current_app.logger.info(
+                Summary(span, created_by, logger).persist(line_summary_client)
+        logger.info(
             (
                 f"Finish writing trace to cosmosdb, total spans count: {len(all_spans)}."
                 f" Duration {datetime.now() - start_time}."
@@ -108,5 +155,5 @@ def _try_write_trace_to_cosmosdb(all_spans):
 
     except Exception as e:
         stack_trace = traceback.format_exc()
-        current_app.logger.error(f"Failed to write trace to cosmosdb: {e}, stack trace is {stack_trace}")
+        logger.error(f"Failed to write trace to cosmosdb: {e}, stack trace is {stack_trace}")
         return