+++ src/promptflow/CHANGELOG.md
@@ -13,6 +13,11 @@
 
 - [SDK/CLI] Support `AzureOpenAIConnection.from_env` and `OpenAIConnection.from_env`. Reach more details [here](https://microsoft.github.io/promptflow/how-to-guides/manage-connections.html#load-from-environment-variables).
 
+### Bugs Fixed
+
+- [SDK/CLI] environment variable `PF_HOME_DIRECTORY` doesn't work for run details & logs.
+- [SDK] `connection.provider` config doesn't work when calling flow as a function.
+
 ## 1.6.0 (2024.03.01)
 
 ### Features Added